Modified colocalisation analysis
The previous analysis considered both up and down regulated genes and
thus introduced a systematic bias which led to the wrong conclusion. For
now, we fix this by using only upregulated genes (down would work just
as well). We might revisit this later to include all genes in an
unbiased manner.