Corrected tRNA expression counts by acceptor/type
We previously used the mean to calculate the per-acceptor and per-type
expression of tRNA genes. We now use the sum instead, since that yields
more robust correlation coefficients. I’m not sure the mean is a valid
measure here, but I cannot discern why it makes a difference at all.