Analyse codon/anticodon correlation with wobbling
First attempt at wobble pairing, which uses the data from the dos Reis
table. Since it turns out that this approach makes many assumptions
about the distribution of codon/anticodon pairing which are not
corroborated by data, we will subsequently attempt a different approach.

This is its own file rather than a separate branch since we want to make
both approaches available to reviewers in the main branch.